ha-guest-entry/requirements.txt
aiohttp>=3.9
aiofiles>=23.2
bcrypt>=4.1
pyjwt>=2.8
pyyaml>=6.0
